chore: bump devframe to 0.1.22, adopt renamed APIs
v0.1.22 renamed the `Devtool*` surface to `Devframe*`. Migrate off the
deprecated APIs we use:

- `defineDevtool` → `defineDevframe`
- `connectDevtool` → `connectDevframe`
- Drop the now-redundant `DevtoolDefinition` annotation (inferred from
  `defineDevframe`).

Also rename `src/devtool.ts` → `src/devframe.ts` and the default export
`devtool` → `devframe` to match the new template convention (matches
the `src/devframe-vite-plugin.ts` filename already alongside it).
`DevtoolFlags` is project-internal naming and stays.
