+++ pnpm-workspace.yaml
@@ -28,7 +28,7 @@ catalog:
   bundle-require: ^5.1.0
   cac: ^7.0.0
   chokidar: ^5.0.0
-  devframe: ^0.1.21
+  devframe: ^0.1.22
   esbuild: ^0.28.0
   eslint: ^10.3.0
   find-up: ^8.0.0
